Increase sleep time to 1s avaiting connection
If AIC is not running we need to have some sleep interval
retrying connection. To avoid polling with high CPU usage
and/or thrashing logs we will be doing that each 1 second.

AIC might not be running in 2 cases:
1. Due to wrong start sequency
2. If AIC went down and requires restart

Keep in mind 1 second connection sleep interval which needs
to be counted towards connection latency.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>